Refactor: extract peerlist fns to new file
