+++ prog/kube-utils/annotations.go
@@ -2,13 +2,10 @@
 In order to keep track of active weave peers, we use annotations on the Kubernetes cluster.
 
 Kubernetes uses etcd to distribute and synchronise these annotations so we don't have to.
-
-This module deals with operations on the peerlist backed by Kubernetes' annotation mechanism.
 */
 package main
 
 import (
-	"encoding/json"
 	"log"
 	"time"
 
@@ -34,38 +31,6 @@ func newConfigMapAnnotations(ns string, configMapName string, clientset *kuberne
 		Namespace:     ns,
 		ConfigMapName: configMapName,
 		Client:        clientset.CoreV1(),
-	}
-}
-
-type peerList struct {
-	Peers []peerInfo
-}
-
-type peerInfo struct {
-	PeerName string // Weave internal unique ID
-	NodeName string // Kubernetes node name
-}
-
-func (pl peerList) contains(peerName string) bool {
-	for _, peer := range pl.Peers {
-		if peer.PeerName == peerName {
-			return true
-		}
-	}
-	return false
-}
-
-func (pl *peerList) add(peerName string, name string) {
-	pl.Peers = append(pl.Peers, peerInfo{PeerName: peerName, NodeName: name})
-}
-
-func (pl *peerList) remove(peerNameToRemove string) {
-	for i := 0; i < len(pl.Peers); {
-		if pl.Peers[i].PeerName == peerNameToRemove {
-			pl.Peers = append(pl.Peers[:i], pl.Peers[i+1:]...)
-		} else {
-			i++
-		}
 	}
 }
 
@@ -109,27 +74,6 @@ func (cml *configMapAnnotations) Init() error {
 		cml.cm.Annotations = make(map[string]string)
 	}
 	return nil
-}
-
-func (cml *configMapAnnotations) GetPeerList() (*peerList, error) {
-	var record peerList
-	if cml.cm == nil {
-		return nil, errors.New("endpoint not initialized, call Init first")
-	}
-	if recordBytes, found := cml.cm.Annotations[KubePeersAnnotationKey]; found {
-		if err := json.Unmarshal([]byte(recordBytes), &record); err != nil {
-			return nil, err
-		}
-	}
-	return &record, nil
-}
-
-func (cml *configMapAnnotations) UpdatePeerList(list peerList) error {
-	recordBytes, err := json.Marshal(list)
-	if err != nil {
-		return err
-	}
-	return cml.UpdateAnnotation(KubePeersAnnotationKey, string(recordBytes))
 }
 
 // Clean up a string so it meets the Kubernetes requiremements for Annotation keys:
